+++ stack.h
@@ -4,8 +4,8 @@
 #include <stddef.h>
 #include <stdbool.h>
 
-void stack_push(int number);
-int stack_pop(void);
+void stack_push(float number);
+float stack_pop(void);
 size_t stack_size(void);
 void stack_clear(void);
 bool stack_empty(void);
